backend/requirements.txt
flask==2.3.3
gunicorn==21.2.0
flask-cors==4.0.0
transformers==4.36.0
torch==2.1.1
pytest==7.4.3
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
pymysql==1.1.0
werkzeug==2.3.7
numpy==1.26.4
